Moved everything except index.js into the utils directory for now.
